AnalogInput handler was moved
